Move checking of dependencies to its own module
